requirements.txt
altgraph==0.17.4
packaging==24.2
pefile==2023.2.7
pillow==11.1.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.11
pywin32==308
pywin32-ctypes==0.2.3
